Fix toRep on 32-bit architectures
The previous implementation was badly broken on 32-bit architectures, as it
would `peek` a value of type `int32_t` at the Haskell type `Int64` (for
instance, this could cause the C value `-2 ^ 31` to be read as `2 ^ 31` on the
Haskell side). We now read this value at the appropriate type (`Int32` on
32-bit architectures and `Int64` on 64-bit architectures) and extend it as
needed.

Fixes #26.